Move to autotools, create RPM spec file
In order to ease compilation in vairous envirionments, I have
implemented a configure script using GNU autoconf. Since Qt is still
used, I have not removed all use of qmake. Qt's qmake is still called
from the generated Makefile. This may be changed in the future, but
for now this is the most convenient way to handle setup.

I have also created an RPM spec file, openblox.spec, which is used
to generate an RPM which can be installed on distributions such as
Fedora, openSUSE, CentOS and other distributions which use rpm.